requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.23
Flask-SQLAlchemy>=3.1.1
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
Flask-Login>=0.6.3
python-dotenv>=1.0.0
Werkzeug>=3.0.1
plotly>=5.18.0
dash>=2.14.2
pytest>=7.4.3 